tests/ml-model/requirements-test.txt
# Testing requirements for ML Model service
pytest==8.0.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
fastapi==0.115.14
uvicorn==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
numpy==2.3.1
pillow==11.3.0
onnxruntime==1.22.0
face-recognition==1.3.0

# Test utilities
factory-boy==3.3.0
faker==22.0.0
